+++ spec/lib/karafka/connection/consumer_spec.rb
@@ -16,32 +16,24 @@ RSpec.describe Karafka::Connection::Consumer do
     let(:event) { double }
     let(:built_controller) { double }
     let(:router) { Karafka::Routing::Router.new(event) }
+    let(:async_scope) { listener }
 
     it 'should use fetch on an each listener and route the request' do
-      expect(listener)
-        .to receive(:fetch)
-        .and_yield(event)
-
       expect(subject)
         .to receive(:listeners)
         .and_return(listeners)
-
-      expect(Karafka::Routing::Router)
-        .to receive(:new)
-        .with(event)
-        .and_return(router)
 
       expect(listener)
         .to receive(:controller)
         .and_return(controller)
         .at_least(:once)
 
-      expect(router)
-        .to receive(:build)
-        .and_return(built_controller)
+      expect(listener)
+        .to receive(:async)
+        .and_return(async_scope)
 
-      expect(built_controller)
-        .to receive(:call)
+      expect(async_scope)
+        .to receive(:fetch_loop)
 
       subject.send(:fetch)
     end
@@ -64,6 +56,33 @@ RSpec.describe Karafka::Connection::Consumer do
 
     it 'should create new listeners based on the controllers' do
       expect(subject.send(:listeners)).to eq [listener]
+    end
+  end
+
+  describe '#message_action' do
+    let(:listener) { double }
+    let(:message_proc) { subject.send(:message_action, listener) }
+    let(:incoming_message) { double }
+    let(:router) { double }
+
+    it 'should be a proc' do
+      expect(message_proc).to be_a Proc
+    end
+
+    it 'should build and route to controller based on message' do
+      expect(listener)
+        .to receive(:controller)
+        .and_return(controller)
+
+      expect(Karafka::Routing::Router)
+        .to receive(:new)
+        .with(incoming_message)
+        .and_return(router)
+
+      expect(router)
+        .to receive_message_chain(:build, :call)
+
+      message_proc.call(incoming_message)
     end
   end
 end